+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * `TrailingBlankLines` cop does auto-correction. ([@jonas054][])
 * `FinalNewline` cop does auto-correction. ([@jonas054][])
 * New cop `CyclomaticComplexity` checks the cyclomatic complexity of methods against a configurable max value. ([@jonas054][])
+* [#594](https://github.com/bbatsov/rubocop/pull/594): New parameter `EnforcedStyleForEmptyBraces` with values `space` and `no_space` (default) added to `SpaceAroundBlockBraces`. ([@jonas054][])
 
 ### Changes